+++ error_helper_test.go
@@ -5,14 +5,14 @@ import (
 	"github.com/ztrue/tracerr"
 )
 
-func addFrameA(message string) *tracerr.Error {
+func addFrameA(message string) error {
 	return addFrameB(message)
 }
 
-func addFrameB(message string) *tracerr.Error {
+func addFrameB(message string) error {
 	return addFrameC(message)
 }
 
-func addFrameC(message string) *tracerr.Error {
+func addFrameC(message string) error {
 	return tracerr.New(message)
 }